+++ src/vue/modules/assets/asset-explorer/index.vue
@@ -1,70 +1,83 @@
 <template>
   <div class="assets-explorer">
-    <template v-if="!isLoadFailed">
-      <drawer :is-shown.sync="isDrawerShown">
-        <template v-if="isUpdateMode">
-          <template slot="heading">
-            {{ 'assets.update-drawer-title' | globalize }}
-          </template>
-
-          <update-asset-form-simplified-module
-            :asset-code="selectedBalance.asset.code"
-            @submitted="(isDrawerShown = false) || loadAccountBalances()"
+    <template v-if="isLoaded && accountBalances.length">
+      <div class="app__card-list">
+        <div
+          class="app__card-list-item"
+          v-for="accountBalance in accountBalances"
+          :key="accountBalance.id"
+        >
+          <asset-card
+            :balance="accountBalance"
+            @transfer="transfer"
+            @vue-details="selectBalance(accountBalance)"
           />
-        </template>
-
-        <template v-else>
-          <template slot="heading">
-            {{ 'assets.details-drawer-title' | globalize }}
-          </template>
-
-          <asset-attributes-viewer
-            :asset="selectedBalance.asset"
-          />
-
-          <div class="assets-explorer__actions">
-            <asset-actions
-              :asset="selectedBalance.asset"
-              @update-asset="isUpdateMode = true"
-              @asset-transfered="loadAccountBalances()"
-              @asset-deleted="(isDrawerShown = false) || loadAccountBalances()"
-            />
-          </div>
-        </template>
-      </drawer>
-
-      <div class="assets-explorer__asset-list-wrp">
-        <div class="assets-explorer__asset-list">
-          <template v-for="accountBalance in accountBalances">
-            <card-viewer
-              :asset="accountBalance.asset"
-              :balance="accountBalance"
-              :key="accountBalance.id"
-              @click="selectBalance(accountBalance)"
-            />
-          </template>
-          <template v-for="index in itemsPerSkeletonLoader">
-            <asset-skeleton-loader
-              v-if="!isLoaded && !accountBalances.length"
-              :key="index"
-            />
-          </template>
         </div>
-
-        <no-data-message
-          v-if="isLoaded && !accountBalances.length"
-          icon-name="trending-up"
-          :title="'assets.no-assets-title' | globalize"
-          :message="'assets.no-assets-msg' | globalize"
-        />
       </div>
     </template>
 
-    <template v-if="isLoadFailed">
+    <no-data-message
+      v-else-if="isLoaded && !accountBalances.length"
+      icon-name="trending-up"
+      :title="'assets.no-assets-title' | globalize"
+      :message="'assets.no-assets-msg' | globalize"
+    />
+
+    <template v-else-if="isLoadFailed">
       <p class="assets-explorer__error-msg">
         {{ 'assets.loading-error-msg' | globalize }}
       </p>
     </template>
+
+    <template v-else>
+      <skeleton-cards-loader />
+    </template>
+
+    <drawer :is-shown.sync="isAssetDetailsDrawerShown">
+      <template slot="heading">
+        {{ 'assets.details-drawer-title' | globalize }}
+      </template>
+
+      <asset-attributes-viewer
+        :asset="selectedBalance.asset"
+      />
+
+      <div class="assets-explorer__actions">
+        <asset-actions
+          :asset="selectedBalance.asset"
+          @transfer="transfer"
+          @update-asset="isAssetUpdateDrawerShown = true"
+          @asset-deleted="(isAssetDetailsDrawerShown = false) ||
+            loadAccountBalances()"
+        />
+      </div>
+    </drawer>
+
+    <drawer :is-shown.sync="isAssetUpdateDrawerShown">
+      <template slot="heading">
+        {{ 'assets.update-drawer-title' | globalize }}
+      </template>
+
+      <update-asset-form-simplified-module
+        v-if="selectedBalance.asset"
+        :asset-code="selectedBalance.asset.code"
+        @submitted="(isAssetUpdateDrawerShown = false) ||
+          loadAccountBalances()"
+      />
+    </drawer>
+
+    <drawer :is-shown.sync="isTransferDrawerShown">
+      <template slot="heading">
+        {{ 'transfer-form.form-heading' | globalize }}
+      </template>
+      <transfer-form
+        v-if="selectedBalance.asset"
+        @operation-submitted="(isTransferDrawerShown = false) ||
+          loadAccountBalances()
+        "
+        :asset-to-transfer="selectedBalance.asset.code"
+      />
+    </drawer>
   </div>
 </template>
 
@@ -72,28 +85,30 @@
 import Drawer from '@/vue/common/Drawer'
 import NoDataMessage from '@/vue/common/NoDataMessage'
 
-import CardViewer from '../shared/components/card-viewer'
 import AssetAttributesViewer from '../shared/components/asset-attributes-viewer'
 import AssetActions from './components/asset-actions'
-import AssetSkeletonLoader from './components/asset-skeleton-loader'
+import SkeletonCardsLoader from '@/vue/common/skeleton-loader/SkeletonCardsLoader'
 import UpdateAssetFormSimplifiedModule from '@modules/update-asset-form-simplified'
+import AssetCard from './components/asset-card'
+import TransferForm from '@/vue/forms/TransferForm'
+import UpdateList from '@/vue/mixins/update-list.mixin'
 
 import { mapGetters, mapActions } from 'vuex'
 import { vuexTypes } from '@/vuex'
-
+import { Bus } from '@/js/helpers/event-bus'
 import { ErrorHandler } from '@/js/helpers/error-handler'
-import UpdateList from '@/vue/mixins/update-list.mixin'
 
 export default {
   name: 'assets-explorer',
   components: {
     Drawer,
     NoDataMessage,
-    CardViewer,
     AssetAttributesViewer,
     AssetActions,
-    AssetSkeletonLoader,
+    SkeletonCardsLoader,
     UpdateAssetFormSimplifiedModule,
+    AssetCard,
+    TransferForm,
   },
 
   mixins: [UpdateList],
@@ -101,29 +116,34 @@ export default {
   data: _ => ({
     isLoaded: false,
     isLoadFailed: false,
-    isDrawerShown: false,
+    isAssetDetailsDrawerShown: false,
+    isTransferDrawerShown: false,
+    isAssetUpdateDrawerShown: false,
     selectedBalance: {},
-    itemsPerSkeletonLoader: 3,
+    businessOwnerId: '',
     isUpdateMode: false,
   }),
 
   computed: {
     ...mapGetters({
       accountBalancesByOwner: vuexTypes.accountBalancesByOwner,
+      accountBalanceByCode: vuexTypes.accountBalanceByCode,
       accountOwnedAssetsBalances: vuexTypes.accountOwnedAssetsBalances,
-      isBusinessToBrowse: vuexTypes.isBusinessToBrowse,
-      businessToBrowse: vuexTypes.businessToBrowse,
+      isAccountGeneral: vuexTypes.isAccountGeneral,
+      myBusinesses: vuexTypes.myBusinesses,
     }),
 
     accountBalances () {
       try {
         let accountBalances = []
-        if (this.isBusinessToBrowse) {
-          // eslint-disable-next-line max-len
-          const accountId = this.$route && this.$route.query && this.$route.query.owner
-            ? this.$route.query.owner
-            : this.businessToBrowse.accountId
-          accountBalances = this.accountBalancesByOwner(accountId)
+        if (this.isAccountGeneral) {
+          /* eslint-disable max-len */
+          let businessAccountBalances = this.businessOwnerId
+            ? this.accountBalancesByOwner(this.businessOwnerId)
+            : this.myBusinesses.flatMap(business => this.accountBalancesByOwner(business.accountId))
+          /* eslint-enable max-len */
+
+          accountBalances = businessAccountBalances
             .filter(item => +item.balance > 0)
         } else {
           accountBalances = this.accountOwnedAssetsBalances
@@ -135,6 +155,11 @@ export default {
         return []
       }
     },
+    isDrawerShown () {
+      return this.isAssetDetailsDrawerShown ||
+         this.isTransferDrawerShown ||
+        this.isAssetUpdateDrawerShown
+    },
   },
 
   async created () {
@@ -145,6 +170,10 @@ export default {
       this.isLoadFailed = true
       ErrorHandler.processWithoutFeedback()
     }
+
+    Bus.on('assets:setBusinessOwnerId', id => {
+      this.businessOwnerId = id || ''
+    })
 
     this.listenUpdateList('assets:updateList', this.loadAccountBalancesAndSetSelectedBalance)
   },
@@ -162,7 +191,12 @@ export default {
     selectBalance (balance) {
       this.selectedBalance = balance
       this.isUpdateMode = false
-      this.isDrawerShown = true
+      this.isAssetDetailsDrawerShown = true
+    },
+
+    transfer (assetCode) {
+      this.selectedBalance = this.accountBalanceByCode(assetCode)
+      this.isTransferDrawerShown = true
     },
 
     async loadAccountBalancesAndSetSelectedBalance () {
@@ -181,7 +215,6 @@ export default {
 <style lang="scss" scoped>
 @import '~@scss/mixins';
 
-$asset-card-margin: 0.75rem;
 $media-small-height: 460px;
 
 .assets-explorer__actions {
@@ -190,12 +223,5 @@ $media-small-height: 460px;
   @include respond-to-height($media-small-height) {
     margin-top: 2.4rem;
   }
-}
-
-.assets-explorer__asset-list {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: flex-start;
-  margin: -$asset-card-margin;
 }
 </style>