+++ src/vue/pages/Business.vue
@@ -60,7 +60,7 @@
 
 <script>
 import AppFooter from '@/vue/navigation/Footer'
-import BusinessDescription from '@/vue/pages/business/BusinessDescription'
+import BusinessDescription from '@/vue/pages/business-viewer/BusinessDescription'
 import AtomicSwapsExplore from '@/vue/pages/atomic-swaps/AtomicSwapsExplore'
 import NoDataMessage from '@/vue/common/NoDataMessage'
 import Loader from '@/vue/common/Loader'
@@ -79,7 +79,6 @@ export default {
     Loader,
     NoDataMessage,
   },
-  mixins: [],
   props: {
     id: {
       type: String,
@@ -94,12 +93,8 @@ export default {
       isFailed: false,
     }
   },
-  computed: {},
-  watch: {},
   async created () {
     this.getBusiness()
-  },
-  destroyed () {
   },
   methods: {
     async getBusiness () {
@@ -210,6 +205,13 @@ export default {
     }
   }
 
+  .business__industry {
+    text-overflow: ellipsis;
+    overflow: hidden;
+    white-space: nowrap;
+    color: $col-text;
+  }
+
   .business__description {
     padding: 2.4rem;
     margin-bottom: 2.4rem;
@@ -233,12 +235,5 @@ export default {
     margin-top: 2.4rem;
     padding: 0 1.6rem;
     width: 100%;
-  }
-
-  .business__industry {
-    color: $col-text;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
   }
 </style>