+++ src/vue/pages/pay/PaySkeleton.vue
@@ -87,8 +87,6 @@ export default {
 @import '~@scss/variables.scss';
 @import '~@scss/mixins.scss';
 
-$media-small-desktop: 960px;
-
 .pay-skeleton {
   padding: 2.4rem;
   display: flex;