+++ src/vuex/index.js
@@ -8,6 +8,7 @@ import wallet from './wallet.module'
 import kyc from './kyc.module'
 import identities from './identities.module'
 import keyValue from './key-value.module'
+import businesses from './businesses.module'
 import kycRecovery from './kyc-recovery.module'
 import movementsHistory from './movements-history.module'
 import sponsorshipRequests from './sponsorship-requests.module'
@@ -28,15 +29,23 @@ export const rootModule = {
     [vuexTypes.LOG_OUT] ({ commit }) {
       commit(vuexTypes.CLEAR_STATE)
     },
-    async [vuexTypes.LOG_IN] ({ getters, dispatch }, { email, password }) {
+    // eslint-disable-next-line max-len
+    async [vuexTypes.LOG_IN] ({ getters, dispatch, rootGetters }, { email, password }) {
       await dispatch(vuexTypes.LOAD_WALLET, { email, password })
       await dispatch(vuexTypes.LOAD_ACCOUNT, getters[vuexTypes.walletAccountId])
       await dispatch(vuexTypes.LOAD_KV_ENTRIES)
 
       const isKycRecoveryInProgress = getters[vuexTypes.isKycRecoveryInProgress]
+      const isAccountCorporate = getters[vuexTypes.isAccountCorporate]
 
       if (!isKycRecoveryInProgress) {
         await dispatch(vuexTypes.LOAD_KYC)
+      }
+      if (isAccountCorporate) {
+        await dispatch(
+          vuexTypes.LOAD_BUSINESS,
+          rootGetters[vuexTypes.accountId]
+        )
       }
     },
     async [vuexTypes.RESTORE_SESSION] ({ getters, dispatch }) {
@@ -88,6 +97,7 @@ function buildStore () {
       kycRecovery,
       movementsHistory,
       sponsorshipRequests,
+      businesses,
     },
     plugins: [sessionStoragePlugin],
   })